Rename package playlist
