+++ exporter/interface.go
@@ -1,8 +1,8 @@
 package exporter
 
-import "github.com/coaxial/tizinger/playlist"
+import "github.com/coaxial/tizinger/extractor"
 
 // Client defines the interface for an exporter.
 type Client interface {
-	CreatePlaylist(name string, tracks []playlist.Tracklist) (ok bool, err error)
+	CreatePlaylist(name string, tracks []extractor.Tracklist) (ok bool, err error)
 }
